+++ commands/daemonize_windows.go
@@ -159,7 +159,7 @@ func (cmd *Daemonize) execute() error {
 
 	if admin != "" {
 		fmt.Println()
-		fmt.Printf("   *** THE DEFAULT '%v' USER PASSWORD IS '%v' ***\n", admin, pwd)
+		fmt.Printf("   *** THE '%v' USER PASSWORD IS '%v' ***\n", admin, pwd)
 		fmt.Println()
 	}