+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## [Unreleased]
+
+### Changed
+1. Displayed the created admin user ID and password at the end of the 'daemonize' output.
+2. Added 'fonts' folder to embedded HTML file system.
+
+
 ## [0.8.0](https://github.com/uhppoted/uhppoted-httpd/releases/tag/v0.8.0) - 2022-07-01
 
 1. Initial release